+++ CHANGELOG.md
@@ -16,6 +16,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### 🐛 Fixed
+
+- `@qiskit/qiskit-sim`: Use Gate.x.name for initial values
+
+
 ## [0.7.1] - 2019-04-23
 
 ### ✏️ Changed