+++ CHANGELOG.md
@@ -16,10 +16,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### ✏️ Changed
+
+- `@qiskit/qiskit-sim`: Add phaseShift util function for phase gates
+
 ### 🐛 Fixed
 
 - `@qiskit/qiskit-sim`: Use Gate.x.name for initial values
-
 
 ## [0.7.1] - 2019-04-23